// src/utils/enums.ts
export enum SubscriptionNameEnum {
    TOP_PLAN = "TOP_PLAN",
    HIGH_LIGHT_PLAN = "HIGH_LIGHT_PLAN",
    PUSH_UP_PLAN = "PUSH_UP_PLAN"
}


export enum RENTED_AS_ENUM {
    HOUR = 'HOUR',
    DAY = "DAY",
    WEEKS = 'WEEKS',
    MONTH = "MONTH",
    YEAR = 'YEAR'
}

export enum ORDER_PROCESS_STATUS_ENUM {
    ORDERED = "ORDERED",
    DELIVERED = "DELIVERED",
    RETURNED = "RETURNED"
}

export enum PAYMENT_STATUS_ENUM {
    PAID = "PAID",
    REFUNDED = "REFUNDED",
    PENDING = "PENDING"
}

export enum ORDER_STATUS_ENUM {
    COMPLETED = "COMPLETED",
    REJECTED = "REJECTED",
    CANCELLED = "CANCELLED",
    PENDING = "PENDING"
}

export enum PAYOUT_TYPE_ENUM {
    SECURITY_AMOUNT = "SECURITY_FEE",
    ORDER_AMOUNT = "ORDER_AMOUNT"
}

export enum PAYOUT_PAYMENT_STATUS_ENUM {
    PAID = "PAID",
    REJECTED = "REJECTED",
    PENDING = "PENDING"
}

// *** NOTIFICATION ENUM ***
export enum NOTIFICATION_TYPES_ENUM {
    INFORMATION = "INFO",
    ERROR = "ERROR",
    SUCCESS = "SUCCESS",
    OTHER = 'OTHER'
}
// *** NOTIFICATION ENUM ***